+++ src/apps/components/shell/Shell/Taskbar/StatusArea.svelte
@@ -5,12 +5,12 @@
 
   const { process }: { process: ShellRuntime } = $props();
   const { stackBusy } = process;
-  const { battery } = process.userDaemon?.status! || {}!;
+  const { battery } = process.userDaemon?.power! || {}!;
 </script>
 
 <div class="status-area">
   <StackBusy {stackBusy} />
-  {#if process?.userDaemon?.status}
+  {#if process?.userDaemon?.power}
     <Battery {battery} userPreferences={process.userPreferences} />
   {:else}
     <span class="lucide error-text battery icon-triangle-alert" title="ERR_NO_DAEMON"></span>``
