+++ src/apps/core/initialsetup/runtime.ts
@@ -339,8 +339,8 @@ export class InitialSetupRuntime extends AppProcess {
     );
 
     await this.userDaemon?.account?.getUserInfo();
-    await this.userDaemon?.preferencesCtx?.startPreferencesSync();
-    await this.userDaemon?.files?.startFilesystemSupplier();
+    await this.userDaemon?.init?.startPreferencesSync();
+    await this.userDaemon?.init?.startFilesystemSupplier();
     this.userDaemon?.preferences.update((v) => {
       v.account.displayName = this.displayName();
