+++ src/ts/server/user/daemon/contexts/checks.ts
@@ -1,9 +1,6 @@
-import SafeModeNotice from "$lib/Daemon/SafeModeNotice.svelte";
-import type { AppProcess } from "$ts/apps/process";
 import { MessageBox } from "$ts/dialog";
 import type { DistributionServiceProcess } from "$ts/distrib";
 import { StoreItemIcon } from "$ts/distrib/util";
-import { KernelStateHandler } from "$ts/getters";
 import { NightlyLogo } from "$ts/images/branding";
 import { ArcBuild } from "$ts/metadata/build";
 import type { MessagingInterface } from "$ts/server/messaging";
@@ -13,19 +10,9 @@ import { UserContext } from "../context";
 
 export class ChecksUserContext extends UserContext {
   public NIGHTLY = false;
-  private registeredAnchors: HTMLAnchorElement[] = [];
-  public anchorInterceptObserver?: MutationObserver;
 
   constructor(id: string, daemon: UserDaemon) {
     super(id, daemon);
-  }
-
-  async _init() {
-    this.startAnchorRedirectionIntercept();
-  }
-
-  async _deactivate() {
-    this.anchorInterceptObserver?.disconnect();
   }
 
   checkReducedMotion() {
@@ -153,13 +140,6 @@ export class ChecksUserContext extends UserContext {
       });
     }
   }
-  async waitForLeaveInvocationAllow() {
-    return new Promise<void>((r) => {
-      const interval = setInterval(() => {
-        if (!this.daemon._blockLeaveInvocations) r(clearInterval(interval));
-      }, 1);
-    });
-  }
 
   checkNightly() {
     const isNightly = !!this.env.get(`NIGHTLY_WHODIS_${ArcBuild()}`);
@@ -182,118 +162,5 @@ export class ChecksUserContext extends UserContext {
       this.pid
     );
     this.NIGHTLY = true;
-  }
-
-  safeModeNotice() {
-    MessageBox(
-      {
-        title: "ArcOS is running in safe mode",
-        content: SafeModeNotice,
-        image: "WarningIcon",
-        sound: "arcos.dialog.warning",
-        buttons: [
-          { caption: "Restart now", action: () => this.daemon.power?.restart() },
-          { caption: "Okay", action: () => {}, suggested: true },
-        ],
-      },
-      +this.env.get("shell_pid"),
-      true
-    );
-  }
-
-  iHaveFeedback(process: AppProcess) {
-    this.daemon.spawn?.spawnApp(
-      "BugHuntCreator",
-      undefined,
-      `[${process.app.id}] Feedback report - ${process.windowTitle()}`,
-      `Thank you for submitting feedback to ArcOS! Any feedback is of great help to make ArcOS the best I can. Please be so kind and fill out the following form:
-  
-  1. Do you want to submit a new 'app', 'feature', or 'other'? Please answer one.
-     - Your answer:
-  
-  2. What do you want me to implement?
-     - Your answer:
-  
-  3. How should I go about this? Any ideas?
-     - Your answer:
-  
-  4. Did a previous version of ArcOS include this (v5 or v6)?
-     - Your answer:
-  
-  5. Convince me why I should implement this feature.
-     - Your answer:
-  
-  
-  **Do not change any of the information below this line.**
-  
-  ------
-  
-  - Username: ${this.userInfo?.username}
-  - User ID: ${this.userInfo?._id}
-  
-  ------
-  
-  
-  # DISCLAIMER
-  
-  The information provided in this report is subject for review by me or another ArcOS acquaintance. We may contact you using the ArcOS Messages app if we have any additional questions. It's also possible that the feedback you've provided will be converted into a GitHub issue for communication with other developers. By submitting this feedback, you agree to that. The issue will not contain any personal information, any personal information will be filtered out by a human being.`,
-      {
-        sendAnonymously: true,
-        excludeLogs: true,
-        makePublic: true,
-      }
-    );
-  }
-
-  startAnchorRedirectionIntercept() {
-    this.Log("Starting anchor redirection intercept");
-
-    const handle = () => {
-      if (this._disposed) return;
-
-      const anchors = document.querySelectorAll("a");
-
-      for (const anchor of anchors) {
-        const href = anchor.getAttribute("href");
-
-        if (this.registeredAnchors.includes(anchor) || href?.startsWith("@client/")) continue;
-
-        this.registeredAnchors.push(anchor);
-
-        anchor.addEventListener("click", (e) => {
-          const currentState = KernelStateHandler()?.currentState;
-
-          e.preventDefault();
-
-          if (currentState !== "desktop") return;
-
-          MessageBox(
-            {
-              title: "Open this page?",
-              message: `You're about to leave ArcOS to navigate to <code>${anchor.href}</code> in a <b>new tab</b>. Are you sure you want to continue?`,
-              buttons: [
-                {
-                  caption: "Stay here",
-                  action() {},
-                },
-                {
-                  caption: "Proceed",
-                  action() {
-                    window.open(anchor.href, "_blank");
-                  },
-                  suggested: true,
-                },
-              ],
-              image: "GlobeIcon",
-            },
-            +this.env.get("shell_pid"),
-            true
-          );
-        });
-      }
-    };
-
-    this.anchorInterceptObserver = new MutationObserver(handle);
-    this.anchorInterceptObserver.observe(document.body, { childList: true, subtree: true });
   }
 }
